+++ designer/client/api/designerApi.ts
@@ -22,7 +22,7 @@ export class DesignerApi {
 
   async fetchData(id: string) {
     try {
-      const response = await window.fetch(`${id}/api/data`);
+      const response = await window.fetch(`/api/${id}/data`);
       return response.json();
     } catch (e) {
       console.error(e);
