+++ designer/package.json
@@ -51,6 +51,7 @@
     "moment-timezone": "^0.5.31",
     "nanoid": "^3.1.12",
     "react-helmet": "^6.1.0",
+    "react-router-dom": "^5.2.0",
     "resolve": "^1.19.0",
     "schmervice": "^1.5.0",
     "vision": "5.4.3"
